Ok, you can edit polls now (hope it all works right).
An option to delete polls will come soon.

git-svn-id: http://userblogmod.googlecode.com/svn/trunk@165 a294dfe0-ea36-0410-9d4a-05498620a423

+++ root/blog/includes/functions_posting.php
@@ -82,8 +82,6 @@ function handle_basic_posting_data($check = false, $page = 'blog', $mode = 'add'
 				'CATEGORY_LIST'				=> $category_list,
 
 				'S_CAT_0_SELECTED'			=> ((is_array($category) && in_array(0, $category)) || $category === 0),
-				'S_SHOW_CATEGORY_BOX'		=> ($category_list) ? true : false,
-				'S_SHOW_PERMISSIONS_BOX'	=> true,
 				'S_SHOW_POLL_BOX'			=> true,
 			));
 		}
@@ -379,7 +377,7 @@ function submit_blog_poll($poll, $blog_id, $mode = 'add')
 	global $db;
 
 	// Update Poll Tables
-	if (isset($poll['poll_options']) && !empty($poll['poll_options']))
+	if (isset($poll['poll_options']))
 	{
 		$cur_poll_options = array();
 
@@ -399,48 +397,39 @@ function submit_blog_poll($poll, $blog_id, $mode = 'add')
 			$db->sql_freeresult($result);
 		}
 
+		// If edited, we would need to reset votes (since options can be re-ordered above, you can't be sure if the change is for changing the text or adding an option
+		if ($mode == 'edit' && sizeof($poll['poll_options']) != sizeof($cur_poll_options))
+		{
+
+			$sql = 'DELETE FROM ' . BLOGS_POLL_OPTIONS_TABLE . '
+				WHERE blog_id = ' . $blog_id;
+			$db->sql_query($sql);
+			$db->sql_query('DELETE FROM ' . BLOGS_POLL_VOTES_TABLE . ' WHERE blog_id = ' . $blog_id);
+			$db->sql_query('UPDATE ' . BLOGS_POLL_OPTIONS_TABLE . ' SET poll_option_total = 0 WHERE blog_id = ' . $blog_id);
+		}
+
 		$sql_insert_ary = array();
-		
+
 		for ($i = 0, $size = sizeof($poll['poll_options']); $i < $size; $i++)
 		{
 			if (strlen(trim($poll['poll_options'][$i])))
 			{
-				if (empty($cur_poll_options[$i]))
-				{
-					// If we add options we need to put them to the end to be able to preserve votes...
-					$sql_insert_ary[] = array(
-						'poll_option_id'	=> (int) sizeof($cur_poll_options) + 1 + sizeof($sql_insert_ary),
-						'blog_id'			=> (int) $blog_id,
-						'poll_option_text'	=> (string) $poll['poll_options'][$i]
-					);
-				}
-				else if ($poll['poll_options'][$i] != $cur_poll_options[$i])
-				{
-					$sql = 'UPDATE ' . BLOGS_POLL_OPTIONS_TABLE . "
-						SET poll_option_text = '" . $db->sql_escape($poll['poll_options'][$i]) . "'
-						WHERE poll_option_id = " . $cur_poll_options[$i]['poll_option_id'] . '
-							AND blog_id = ' . $blog_id;
-					$db->sql_query($sql);
-				}
+				// If we add options we need to put them to the end to be able to preserve votes...
+				$sql_insert_ary[] = array(
+					'poll_option_id'	=> (int) sizeof($cur_poll_options) + 1 + sizeof($sql_insert_ary),
+					'blog_id'			=> (int) $blog_id,
+					'poll_option_text'	=> (string) $poll['poll_options'][$i]
+				);
 			}
 		}
 
 		$db->sql_multi_insert(BLOGS_POLL_OPTIONS_TABLE, $sql_insert_ary);
-
-		if (sizeof($poll['poll_options']) < sizeof($cur_poll_options))
-		{
-			$sql = 'DELETE FROM ' . BLOGS_POLL_OPTIONS_TABLE . '
-				WHERE poll_option_id > ' . sizeof($poll['poll_options']) . '
-					AND blog_id = ' . $blog_id;
-			$db->sql_query($sql);
-		}
-
-		// If edited, we would need to reset votes (since options can be re-ordered above, you can't be sure if the change is for changing the text or adding an option
-		if ($mode == 'edit' && sizeof($poll['poll_options']) != sizeof($cur_poll_options))
-		{
-			$db->sql_query('DELETE FROM ' . BLOGS_POLL_VOTES_TABLE . ' WHERE blog_id = ' . $blog_id);
-			$db->sql_query('UPDATE ' . BLOGS_POLL_OPTIONS_TABLE . ' SET poll_option_total = 0 WHERE blog_id = ' . $blog_id);
-		}
+	}
+	else if ($mode == 'edit')
+	{
+		$sql = 'DELETE FROM ' . BLOGS_POLL_OPTIONS_TABLE . '
+			WHERE blog_id = ' . $blog_id;
+		$db->sql_query($sql);
 	}
 }
